+++ dev/tests/spec/match.spec.js
@@ -8,15 +8,18 @@ var crossroads = crossroads || require('crossroads');
 describe('Match', function(){
 
     var _prevTypecast;
+    var _prevCase;
 
     beforeEach(function(){
         _prevTypecast = crossroads.shouldTypecast;
+        _prevCase = crossroads.ignoreCase;
     });
 
     afterEach(function(){
         crossroads.removeAllRoutes();
         crossroads.resetState();
         crossroads.shouldTypecast = _prevTypecast;
+        crossroads.ignoreCase = _prevCase;
     });
 
 
@@ -94,6 +97,26 @@ describe('Match', function(){
         expect( s.match('/123/asd/45') ).toBe( true );
         expect( s.match('/123/asd/45/') ).toBe( true );
         expect( s.match('/123/asd/45/qwe') ).toBe( false );
+    });
+
+    it('should not be case sensitive by default', function () {
+        var s = crossroads.addRoute('foo/bar');
+        expect( s.match('foo') ).toBe( false );
+        expect( s.match('Foo') ).toBe( false );
+        expect( s.match('foo/bar') ).toBe( true );
+        expect( s.match('Foo/Bar') ).toBe( true );
+        expect( s.match('FoO/BAR') ).toBe( true );
+    });
+
+    it('should be allow toggling case sensitivity', function () {
+        crossroads.ignoreCase = true;
+
+        var s = crossroads.addRoute('foo/bar');
+        expect( s.match('foo') ).toBe( false );
+        expect( s.match('Foo') ).toBe( false );
+        expect( s.match('foo/bar') ).toBe( true );
+        expect( s.match('Foo/Bar') ).toBe( true );
+        expect( s.match('FoO/BAR') ).toBe( true );
     });
 
     describe('rest params', function () {
@@ -445,16 +468,36 @@ describe('Match', function(){
         describe('basic rules', function(){
 
             it('should allow array options', function(){
+                var s = crossroads.addRoute('/{foo}/{bar}');
+                s.rules = {
+                    foo : ['lorem-ipsum', '123'],
+                    bar : ['DoLoR', '45']
+                };
+
+                expect( s.match('/lorem-ipsum') ).toBe( false );
+                expect( s.match('/lorem-ipsum/DoLoR') ).toBe( true );
+                expect( s.match('/LoReM-IpSuM/DOLoR') ).toBe( true );
+                expect( s.match('lorem-ipsum') ).toBe( false );
+                expect( s.match('/123') ).toBe( false );
+                expect( s.match('123') ).toBe( false );
+                expect( s.match('/123/123') ).toBe( false );
+                expect( s.match('/123/45') ).toBe( true );
+            });
+
+            it('should change array validation behavior when ignoreCase is false', function(){
+                crossroads.ignoreCase = false;
 
                 var s = crossroads.addRoute('/{foo}/{bar}');
 
                 s.rules = {
                     foo : ['lorem-ipsum', '123'],
-                    bar : ['dolor', '45']
+                    bar : ['DoLoR', '45']
                 };
 
                 expect( s.match('/lorem-ipsum') ).toBe( false );
-                expect( s.match('/lorem-ipsum/dolor') ).toBe( true );
+                expect( s.match('/lorem-ipsum/dolor') ).toBe( false );
+                expect( s.match('/lorem-ipsum/DoLoR') ).toBe( true );
+                expect( s.match('/LoReM-IpSuM/DOLoR') ).toBe( false );
                 expect( s.match('lorem-ipsum') ).toBe( false );
                 expect( s.match('/123') ).toBe( false );
                 expect( s.match('123') ).toBe( false );
@@ -462,6 +505,7 @@ describe('Match', function(){
                 expect( s.match('/123/45') ).toBe( true );
 
             });
+
 
             it('should allow RegExp options', function(){
                 var s = crossroads.addRoute('/{foo}/{bar}');